requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==36.0.2
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
idna==3.3
ipykernel==6.9.1
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.2
jupyter-core==4.9.2
nest-asyncio==1.5.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.0.4
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==22.3.0
requests==2.27.1
setuptools==62.1.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
twilio==7.9.0
urllib3==1.26.9
wcwidth==0.2.5
wheel==0.37.1
